Don't vectorize popcount for short vectors
Avoid firing up the AVX-512 steam engines for short vectors.
Just looking at the groundhog shadow of a 512 bits wide AVX
instruction is seemingly enough to send a Xeon CPU from the
baseline power license level 0 into a frequency throttled power
license level of 1. This is much less of a problem on Ice Lake
microarchitectures and beyond, but still measurable in practice.