Add experimental streaming JSONL output to the Document V1 API handler
Adds a streaming JSONL response writer where puts and removes are
written in the Vespa JSONL feed format. Since the output also contains
other non-put/remove entries, this is a strict _superset_ of the feed
format, and clients must take that into account.

One major difference from the "restful" JSON responses (streaming or
not) is that continuation tokens are emitted inline as they are updated
by the visitor session, rather than being emitted once as part of the
response epilogue. This means a client can continuously observe updated
token values and simply remember the last one they have observed.
Visiting can then restart from that point with minimal loss of progress.

The actual structure/contents of the emitted JSONL lines are not yet
set in stone, and this is therefore currently to be considered an
entirely experimental feature that should not be used in production.
Correspondingly, enabling this output format currently requires the
use of an undocumented HTTP parameter, which will **stop working** once
the JSONL output format is set in stone.