Fix vespamalloc edge case in `realloc`
Code did not properly constrain the amount of memory copied across
old and new allocations if the following held:

 1. the old allocation was at least as large as the limit set by
    `mallopt(M_MMAP_THRESHOLD)`, which means it got its own dedicated
    "off-heap" anomymous memory mapped buffer.
 2. the new allocation size is _smaller_ than the old allocation size.

This could lead to heap overwrites or, with a likelihood increasing
with the size of the old buffer, triggering a `SIGSEGV` due to
touching memory outside mapped arena blocks.

This commit ensures that memory copied is capped to the _minimum_
of the old and new allocation sizes.