Don't make assumptions on how many stack frames can be skipped
This appears to be highly platform specific and depends on the
mood of the compiler optimizations, the linker, and the the
current phase of the moon.

Use `noinline` instead of `noipa` since the latter does not
appear to be properly supported on Clang.

Use Abseil-provided macro to prevent tail call optimizations
of dummy functions in test. This brushes some of the inline
assembly clobbering compiler hacks under the sofa.