Add test malloc state explorer emitting mimalloc-internal stats
Since the malloc explorer looks at the presence of a particular symbol
to deduce if mimalloc is present, and to then invoke it, override
this symbol in the test. This effectively creates a mimalloc mock
for statistics. Since it's possible that the test itself is run
under mimalloc, delegate to the _real_ implementation iff it is
present and the test code is not running.