Support range predicates in filter in grouping
This commit contains the frontend for range predicate in filter in
grouping. Adds RangePredicate and extends the GroupingParser. Converts
RangePredicate to RangePredicateNode and allocates identifiable id 178
for range predicate node.

We already support filtering on regex in grouping. This works great for
text fields, but with numeric fields, we probably want filters based on
comparison, and a good way to support this is using ranges.