Rewrite `hwaccelerated` benchmarks to use Google Benchmark
This improves the resolution and accuracy of runtime sampling
and makes it trivial to run each benchmarked function on a
range of input vector lengths (this had to be done manually
prior).

The benchmark suite is dynamically generated based on the set
of targets that are available at compile+runtime. Standard
Google Benchmark command line args/env vars can be used to select
a subset of tests to run, or to export test results as JSON etc.

Also add a method to `IAccelerated` that reports the name of
the underlying vectorization implementation, as the target name
alone may be ambiguous across distinct implementations built
for the same target (e.g. "NEON", "AVX3" etc.).