Explicitly exclude ARM SVE/SVE2 targets for now
SVE/SVE2 is not a strict superset of NEON, which means that
certain very useful 128-bit NEON(_BF16) vector instructions
are _not_ present as "sizeless" SVE vector operations.

In particular:
 - int8 squared Euclidean distance:
   NEON has `ssub` signed subtraction of high/low vector lanes
   with implicit widening. On SVE this needs separate unpack
   high/low instructions followed by a non-widening subtraction,
   increasing instruction count and register pressure.
 - BFloat16 dot product:
   SVE does not have guaranteed BF16 support prior to armv8.6-a
   and its BF16 dot product instruction (BFDOT) does not give the
   same result as NEON BF16 unless `FEAT_EBF16` is present, due
   to differences in rounding behavior. Because of this, dynamic
   target compilation for Highway does not by default use BF16
   instructions for _any_ SVE targets. It is also not clear how
   this could be enabled with today's set of compilation targets,
   as they are not ARM architecture version-oriented.

In practice this means that 128-bit SVE may be _slower_ for some
important operations than 128-bit NEON_BF16. For both the above
ops, the observed relative slowdown is on the order of ~1.5-2x.
The only serious speed increase from SVE is for popcount.

So for now, exclude SVE targets entirely until they've had more
time to cook. If SVE is present, NEON_BF16 is expected to always
be present.

This is based on testing on Google Axion (SVE2_128), Amazon
Graviton 3 (SVE_256) and Amazon Graviton 4 (SVE2_128) nodes, and
will be updated once newer/shinier hardware is available for
testing.