Don't count `TIMEOUT` as a network error signal for statistics reporting
If an operation reaches its deadline in the content node
persistence queues, it will be automatically and immediately
failed out with a `TIMEOUT` response code. This means that
observing this code at a higher level does not necessarily
mean that a network error condition exists; there could just
as well be an overload condition.