Add flexible vector kernel dispatch function tables
With the existing `IAccelerated` inheritance-based approach
to vector kernel dispatch we have the inherent problem that
all function dispatch is "all or nothing" for a given target
(i.e subclass).

However, some targets are _slower_ for certain operations than
targets that are considered _less preferred_. In particular
this affects BF16 kernels on AArch64 SVE/SVE2 due to Highway
not using certain instructions there due to rounding semantics.

The chosen hacky workarounds have been to either a) entirely
disable some targets, or to b) conditionally not compile certain
functions. a) has the immediate downside that it prevents
the use of functions that are _faster_ on the target and b)
has the downside that such functions remain unavailable for
benchmarking and testing.

This commit introduces an entirely new mechanism for dispatching
calls to vector kernels by adding explicit function tables to
the mix. Each target exports a (potentially sparse) function
table and at runtime we build a _composite_ table where we pick
and choose individual functions from the targets available on the
current platform. Each target may tag specific functions as
_suboptimal_, which causes these to not be picked as part of
the composite table compilation.

This means that when using Highway vectorization, we can fall
back to the best auto-vectorized functions for the parts that
are not yet covered by Highway kernels, rather than having to
fall back to the generic implementation (as that's the base
class).

Calling vectorized functions can now be done using freestanding
functions (no need to carry around `IAccelerated` state) which
dispatch via global function pointers. These function pointers
are initialized by the linker to point to valid (but possibly
suboptimal) baseline implementations, and are patched to point
to the functions resolved by the "optimal" composite table
at some point during the global constructor phase. Unit tests
and benchmarks can also replace the global function dispatch
table, as long as this is done in a single-threaded manner.

Eventually we probably want to compile the function table by
using auto-tuning for current host instead of hard-coding what
kernels we believe are (sub-)optimal.

The existing `IAccelerated` interface is left unchanged, and
all other wiring remains as-is for now. Changing this will be
done as a follow-up.