Do not close the `ContainerThreadPool`
Since `DocprocService` no longer owns the thread pool's lifetime (it is now shared across all `DocprocService` instances), we must no longer close the thread pool in the destructor, as this would be destructive to other components that require the same thread pool, possibly across generations. This could explain why the tests that broke when closing the shared thread pool worked for the first generation, and then suddenly stopped working.