Fix function table exclusion logic
Parameter name was ambiguously named so that it could be
interpreted in two separate ways ("ignore functions that
are suboptimal" and "ignore _that_ functions are suboptimal")
and both these interpretations found their way into the
code. Only the former is correct, so fix logic and rename
the variable. Also add explicit testing that of course
should have been present initially.

This bug meant that we'd still use functions for better targets
even though we had tagged them as suboptimal. This does not
affect correctness, as all functions are tested, but a subset
of operations may be slightly slower than they should be.