Ensure we emit continuation tokens in error edge cases when streaming
The Document V1 API handler logic for handling visitor session
errors did not consistently take into account the difference in
semantics between non-streaming and streaming responses. This
could cause `continuation` tokens to _not_ be written to the
response stream in certain error edge cases, because the code
assumed it would be able to trigger an explicit client error
via the HTTP response code. But in streaming mode we've already
sent the response headers, so that communication channel is closed.

Ensure we emit the continuation token (unless complete) in error
scenarios to avoid having it erroneously appear as if the visitor
session has completed (signalled by not having a `continuation`
object in the response).