Handle failure edge during first visitor before any buckets have been completed
If no initial progress token was provided by the client (i.e.
this is the first visit of potentially many) we must remember
the progress token implicitly created by the visitor session
during its bootstrap, as this token will represent the
completely unfinished visit state.

This is because a session failure prior to receiving even a
single bucket will not provide us with an onProgress control
handler callback, nor will the VisitorParameters have a token.
If we then don't remember the session's bootstrap token, we
won't have a token to communicate to the client, and it may
erroneously believe that visiting has fully completed.