Avoid over/underflow for MessageBus pending count/size counters
Overflow is avoided by moving from `uint32_t` to `size_t`.

Underflow is avoided by unconditionally incrementing pending
counters if a message is not busy-bounced, not just if at least
one config limit is non-zero. This is because we already
unconditionally _decrement_ the counters in the reply handling
logic.

These over/underflows would not happen in a default configuration
since content nodes do not have an incoming MessageBus server
and the distributor by default limits on pending count (not size),
meaning that at least one config limit was non-zero and we would
always increment the pending counters.