Track estimated distributor memory usage from put/update operations
This adds a top-level `MemoryUsageTracker` utility that
is shared between all distributor stripes, and a simple RAII
token for individually tracked operations. Put and Update
operations use the wire serialization size as a (lossy) proxy
of the in-memory cost for the operations they send out.
Additionally, update operations that trigger write-repair
will also transitively track the extra cost incurred by the
convergence-forcing implicit Put operation.

For now this memory tracking is only wired to a singular
top-level value metric. Once we verify that the value of
this metric correlates reasonably well with the actual RSS
of the distributor process, it can be utilized for other
purposes.

Since memory usage may fluctuate within the metric sampling
period, explicitly track the maximum observed usage value
within the period, with destructive sampling that resets it
back to 0 when the metric is updated.