Fix top-level distributor memory usage metric aggregation
The distributor has to wrangle metric snapshotting in a very
particular manner to present a unified view of metrics across
all stripes and itself. Due to some confusing naming of the
top level metric set, the memory usage metric was not wired in
a way that it would be properly propagated to snapshots.

Annoyingly, the previous wiring had the expected results for
the _first_ snapshot taken (but not subsequent ones), making
it seem like it worked fine during local testing.

With this commit the wiring should be fixed and metric set
naming has been updated to more accurately reflect semantics.